refactor : BettingResult 패키지 변경
