+++ src/main/java/blackjack/domain/participants/BettingResult.java
@@ -1,7 +1,5 @@
-package blackjack.domain;
+package blackjack.domain.participants;
 
-import blackjack.domain.participants.Player;
-import blackjack.domain.participants.State;
 import java.util.LinkedHashMap;
 import java.util.Map;
